Fix App Author Tipping and Reviews
Wrote a review for the quest "Splice", and wanted also to tip Ben for the quest, but as soon as I hit "submit" for the review 
it exited me from the quest.  In the past, I've tipped then had to remember to go back and submit the review, and the last time I did
that (for the "Spirit of Festivus") the review was not actually sent (though it did exit me from quest). 

Even a well meaning player is unlikely (especially when gaming with friends) to go back and through the trouble of getting to an 
end screen so they can tip if they get booted out after a review. 

We should fix and make sure they can tip before they review, or let them do both from one screen (the more difficult it is, the less likely
people on leisure time are to do it). 

Greg